InteractRank/interactrank/requirements.txt
torchinfo==1.8.0
hydra-core==1.3.1
overrides==7.7.0
simplejson==3.19.2
torchsnapshot==0.1.0
torchrec==0.5.0
fbgemm-gpu==0.8.0
numpy==1.23.4
wandb==0.19.10